Cleanup in prep for XML solib list support on svr4
This commit incorporates 3rd-party changes made in GNU's gdb that are
necessary to cleanly merge the "Add library-list-svr4.dtd support"
commit that follows. These changes were originally made as a
refactorization effort, in preparation for the new library-list-svr4
feature.

This commit is a dependency of the fix for the following bug:
  http://code.google.com/p/android/issues/detail?id=38402

See
  [1] http://www.cygwin.com/ml/gdb-patches/2011-10/msg00057.html
  [2] http://www.cygwin.com/ml/gdb-patches/2011-10/msg00058.html

Change-Id: I33acf5610371946c595398241fc8b373ae9b31b4