Add library-list-svr4.dtd support
This commit incorporates 3rd-party changes made in GNU's gdb, and is
the first of 2 commits addressing the following bug:
  http://code.google.com/p/android/issues/detail?id=38402

Every time GDB gets notification from the linker that the inferior has
loaded a new solib, GDB was fully walking the inferior's entire link
map, gathering all the available information.  This is a very expensive
process, especially over a remote transport such as ADB.  This change
improves performance by walking the link map from within gdbserver,
and then providing all the needed information back to GDB in a
single transaction, formatted as XML.

Similar support already existed, however it was broken in numerous ways
that caused GDB to fall back to the slower path.

With this change, this support will still be broken on Android, however
the next commit will address the remaining, bionic-specific
requirements.

Also includes a minor patch to address a compilation error that occurs
when building GDB with "--without-expat".

See
  [1] http://www.cygwin.com/ml/gdb-patches/2011-11/msg00103.html
  [2] http://sourceware.org/ml/gdb-patches/2011-12/msg00075.html